Test presence of array keys instead of values of selected options in Html::getOptions()
Fix: The $selected parameter array(1 => null) would not select the option with a value attribute "1", despite the key being present.
Now the behavior is as documented (and intended): "The optional $selected parameter may be an array, in which case all IDs found in the array keys are selected.  The arrays' values are ignored."