fix(calendar): ensure exception dates are in correct order (CLX-2839)
This is required as in JavaScript the keys of an array have no fixed order.
Meaning that when we order the indexes of a JavaScript array, the order
might get reset by the engine. Therefore, if we do not explicitly define
our own indexed, but instead let JavaScript index the array on its own,
the set order will be kept in place